requirements.txt
emoji
hatch
Pillow
pytest
transformers[torch]
